CommentPage can load the LinkObject by itself
Previously CommentPage is depends on the LinkModel to get the link
object. This causes the CommentPage can not be push to pageStack outside
LinkModel. Now when the link object is not provided by the LinkModel
(pushed to pageStack outside a LinkModel) CommentModel will parse the
LinkObject from the comments response and expose to QML (by converting
to a QVariantMap). This allow pushing CommentPage to pageStack while in
MessagePage.