sailfish: don't be explicit on the files to be packaged in %{_datadir}
As notification configuration files are not allowed in harbour, we have to make its inclusion conditional.
in yaml it is not possible to use conditionals, so we just include the %{_datadir} in its entirety and let the .pro
file decide on its inclusion. This approach avoids having to keep a diff around in a harbour branch.