sailfish: Rewrite SignInPage with SailfishWebView
The page has been rewritten with a Sailfish WebView element instead of
a SilicaWebview.

To adjust the page, improve the layout and work better with the
Sailfish WebView I replaced the AbstractPage with Dialog.
The page was more of a dialog in any case, the reload option was
redundant.

While doing so I replaced the copyright text with SPDX tags.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>